+++ libs/ngx-material-intl-tel-input-lib/src/index.ts
@@ -1,1 +1,2 @@
 export * from './lib/ngx-material-intl-tel-input-lib/ngx-material-intl-tel-input-lib.component';
+export * from './lib/types/text-labels.type';